refactor: 타입 및 리듀서 코드 분리
