+++ client/src/components/Question/Paragraph/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Container, Text } from "./Paragraph.style";
+import { Container } from "./Paragraph.style";
 
 function Paragraph() {
   return <Container>장문형 텍스트</Container>;
